self-learning/portfolio/requirements.txt
streamlit==1.28.0
streamlit-lottie==0.0.5
streamlit-three==0.1.3
pillow==10.0.0
requests==2.31.0 